// src/interfaces/Restaurant.ts
export default interface Restaurant {
    id: string
    area: string
    url: string
    image_name: string
    name: string
    genre: string
    tel: string
    business_day_info: string
    address: string
    latitude: string
    longitude: string
    comment: string
    takeout_available: number
}